+++ packages/image_picker/image_picker/integration_test/old_image_picker_test.dart
@@ -1,3 +1,5 @@
+// @dart=2.9
+
 import 'package:integration_test/integration_test.dart';
 
 void main() {