+++ packages/package_info/example/android/app/src/main/java/io/flutter/plugins/packageinfoexample/EmbedderV1Activity.java
@@ -5,6 +5,7 @@
 package io.flutter.plugins.packageinfoexample;
 
 import android.os.Bundle;
+import dev.flutter.plugins.integration_test.IntegrationTestPlugin;
 import io.flutter.app.FlutterActivity;
 import io.flutter.plugins.packageinfo.PackageInfoPlugin;
 
@@ -14,5 +15,7 @@ public class EmbedderV1Activity extends FlutterActivity {
     super.onCreate(savedInstanceState);
     PackageInfoPlugin.registerWith(
         registrarFor("io.flutter.plugins.packageinfo.PackageInfoPlugin"));
+    IntegrationTestPlugin.registerWith(
+        registrarFor("dev.flutter.plugins.integration_test.IntegrationTestPlugin"));
   }
 }