+++ packages/path_provider/path_provider_linux/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.0.0
+
+* Migrate to null safety.
+
 ## 0.1.1+3
 
 * Update Flutter SDK constraint.